Merged the tst-hirizarry/l_dev fork  as of today with the master branch of Jeff's form of the Java-WebSocket  library originally from TooTallNate
We need to do a better form of vendor branch including but for now the
source is all in the zip in the APIs directory.

The going theory:

A runtime exception in a WebsocketWorker will stop the whole
WebsocketServer it ultimately causes handleFatal to get called

A runtime exception could happen anywhere during the decoding but it
clearly could happen during the Websocket handshake handled in the Draft
and its subclasses

The fork from tst-hirizarry/l_dev converts IncompleteHandshakeException
to a checked exception and therefore probably solves cases where a
malformed client could inadvertently cause a runtimeexception and bring
down the server.